Replace Inner and InnerPtr with AsRef/AsMut
This replaces custom traits for inner pointers with safe, lifetime-bound access, via standard traits AsRef and AsMut. Of course, the caller will still be forced to use `unsafe` for any FFI calls that operate on those pointers, but the access to references themselves is safe and becomes a bit more convenient via those traits.