Widgets revamp
This adds typed variations of widgets, so that it's impossible to pass incompatible value to a widget set_value, or try to access children on non-grouping widget and so on.

More specifically, this adds `widget::{GroupWidget, TextWidget, RangeWidget, ToggleWidget, RadioWidget, ButtonWidget}` types that each provides its own strongly typed API.

All those types inherit from `WidgetBase` that also provides core properties such as `id`, `name` and so on that can be accessed on any of strongly widget types as well.

There is also a `Widget` enum that is simply a container for arbitrary widget and can be matched against or converted into the specific widget type.

Finally, top-level function `config` was updated to return `GroupWidget` and `config_key` has a generic param that allows to immediately specify which widget type you want (since in most cases, when user provides a hardcoded widget name, they likely know its type as well).

This is fairly big change, but I think the ergonomics of statically typed widgets are worth it.