Make some methods infallible
While technically all functions in libgphoto2 return status code,
a lot of them are simple getters that always return GP_OK unless you
pass null pointer or such.

As long as our bindings are correct, this should never happen in our
Rust code, and if it does, it's an actual bug in our code and can be
better represented by a panic.

For most users, this should make consumption of such getters simpler.