Revamp CameraList
Instead of exposing CameraList directly and allocating Vec unconditionally, expose two iterators - one for actual camera list, another for file list (both of which are represented by CameraList in libgphoto2).

Additionally, camera list item is now represented by a CameraDesc instead of a tuple, which makes it a bit more intuitive thanks to readable field names as welll as makes it easier to pass to context.get_camera().