Add snapshot-based testing
This adds snapshot-based testing based on insta and (almost undocumented) virtual camera driver in libgphoto2, which should make it a lot easier to refactor code without having to verify that it still works with real cameras. These tests don't cover all the methods yet, but they already helped find couple of bugs like #22, #27, as well as few small issues in libgphoto2 itself.

This took a long time and surprisingly lots of effort to integrate (partially because of mentioned upstream issues with the driver, partially because of lack of documentation, and partially because of cross-platform shenanigans), but, on the bright side, it should make things a bit easier from now on, both for gphoto2-rs itself as well as for downstream users, who can now add `gphoto2-test` as a dependency and write their own tests based on it too.